requirements.txt
pytesseract
Pillow
openai
python-dotenv
gTTS